Refactor: backend Dockerfile .lock 파일 처리 수정
- Python 버전 3.13 업그레이드
- 검색조건 수정

Refactor: front refactoring
